+++ greenfield/eldenring/missable_locations.py
@@ -290,4 +290,4 @@ MISSABLE_LOCATIONS = {
     7900274: 'questline',
 }
 
-_GEN_STAMP = {'inputs_hash': 'sha256:8a2c6e75be1dd4ee1023dcc5b234a3f0523a48d2f45a8ce175c20ee92c72c817', 'module': 'missable_locations.py', 'body_sha256': 'sha256:af8454e979c70125c180d074728eec1b000c9dbd29be3bed2c9da9d5e7bb8832'}
+_GEN_STAMP = {'inputs_hash': 'sha256:c08941cc1c41aec8bff7b2581a9253b90b1df3a1b02ff9843da153a7011c6bcd', 'module': 'missable_locations.py', 'body_sha256': 'sha256:af8454e979c70125c180d074728eec1b000c9dbd29be3bed2c9da9d5e7bb8832'}